brcmfmac: Fix transmit stalls caused by bad patch
Commit [1] changes the transmit credits policy to allow two slots for
control traffic. In doing so the author removed some careful (u8) casts
added to cope with wrap-around in an 8-bit sequence number. Removing
the casts causes the appearance of a lack of credit as soon as the
sequence number wraps back to zero.

It's staggering that this was accepted upstream.

Fixes: [1]

Signed-off-by: Phil Elwell <phil@raspberrypi.com>

[1] b41c232d3366 ("brcmfmac: reserve 2 credits for host tx control path")